Use one 'cohorts' database table instead of two, whilst adding a new 'authorisation_statuses' table. This is a significant refactor, and touches several areas including: - Database tables - Models - Enums - Teacher cohort relationships - Seeding - Admin panel pages
The purpose of these changes is for better scalability and flexibility of the codebase in the long run. It reduces the previous repetition on the database level of cohort data which is ultimately the same across two very similar purposes (authorisation cohorts and update cohorts), whilst separating the strict determination of a teacher's authorisation status from their cohorts. In the future this will be expanded upon to allow admin users to change a teacher's authorisation status manually where required.